基本跑通 standard.response.go中存在bug，等待官方解决。

+++ engine/foo.go
@@ -2,5 +2,5 @@ package engine
 
 var (
 	// 文件上传默认内存缓存大小，默认值是 1 << 32 (32MB)。
-	MaxMemory int = 32 << 20
+	MaxMemory int64 = 32 << 20
 )
